feat: Change warmup_delay from minutes (float) to seconds (integer)
- Updated API validation: min:1, max:3600, default:60
- Changed entity model from float64 to int
- Updated task executor to use seconds directly (removed *60 conversion)
- Changed database column from DECIMAL(5,2) to INTEGER
- Updated frontend: step=1, precision=0, labels show "seconds"

+++ core/internal/service/batch_mail/batch_mail.go
@@ -37,8 +37,8 @@ type CreateTaskArgs struct {
 	//Etypes      string `json:"etypes"`      // Email types (e.g., group IDs)
 	Remark    string `json:"remark"`     // Task remark
 	StartTime int    `json:"start_time"` // Scheduled start time
-	Warmup      int     `json:"warmup"`       // Warmup campaign association
-	WarmupDelay float64 `json:"warmup_delay"` // Warmup delay in minutes
+	Warmup      int `json:"warmup"`       // Warmup campaign association
+	WarmupDelay int `json:"warmup_delay"` // Warmup delay in seconds
 	AddType   int    `json:"add_type"`   // Add type (0: normal)
 	GroupId   int    `json:"group_id"`   // Groups to unsubscribe from
 	TagIds    []int  `json:"tag_ids"`    // Tag IDs for filtering contacts
